patch: Reorganize directory structure for Actions, Handlers, and Validators
