+++ README.md
@@ -23,10 +23,10 @@ skeets**
 -   [Overview](#overview)
 -   [Agent](./src/agent/README.md)
 -   [Validators](./src/validations/README.md)
--   [Actions](./src/actions/README.md)
--   [Handlers](./src/handlers/README.md)
-    -   [Record Handlers](./src/handlers/README.md)
-    -   [Pre-made Handlers](./src/handlers/premade-handlers/README.md)
+-   [Actions](src/actions/message-actions/README.md)
+-   [Handlers](src/handlers/message-handlers/README.md)
+    -   [Record Handlers](src/handlers/message-handlers/README.md)
+    -   [Pre-made Handlers](src/handlers/message-handlers/premade-handlers/README.md)
 -   [Jetsteam Firehose Subscription](./src/firehose/README.md)
 -   [Utility Functions](./src/utils/README.md)
 -   [Jetstream Types](./src/types/README.md)
