+++ src/actions/message-actions/AbstractMessageAction.ts
@@ -1,7 +1,7 @@
 import { RepoOp } from '@atproto/api/dist/client/types/com/atproto/sync/subscribeRepos';
-import { HandlerAgent } from '../agent/HandlerAgent';
-import { JetstreamMessage } from '../types/JetstreamTypes';
-import { DebugLog } from '../utils/DebugLog';
+import { HandlerAgent } from '../../agent/HandlerAgent';
+import { JetstreamMessage } from '../../types/JetstreamTypes';
+import { DebugLog } from '../../utils/DebugLog';
 
 export abstract class AbstractMessageAction {
     constructor() {}
