+++ src/actions/message-actions/TestAction.ts
@@ -1,6 +1,6 @@
-import { HandlerAgent } from '../agent/HandlerAgent';
-import { JetstreamMessage } from '../types/JetstreamTypes';
-import { DebugLog } from '../utils/DebugLog';
+import { HandlerAgent } from '../../agent/HandlerAgent';
+import { JetstreamMessage } from '../../types/JetstreamTypes';
+import { DebugLog } from '../../utils/DebugLog';
 import { AbstractMessageAction } from './AbstractMessageAction';
 
 export class TestAction extends AbstractMessageAction {
