+++ src/handlers/message-handlers/premade-handlers/GoodBotHandler.ts
@@ -1,8 +1,8 @@
-import { IsGoodBotValidator } from '../../validations/BotValidators';
-import { DebugLogAction } from '../../actions/LoggingActions';
-import { HandlerAgent } from '../../agent/HandlerAgent';
-import { ReplyToSkeetAction } from '../../actions/post/SkeetActions';
-import { CreateSkeetMessage } from '../../types/JetstreamTypes';
+import { IsGoodBotValidator } from '../../../validations/message-validators/BotValidators';
+import { DebugLogAction } from '../../../actions/message-actions/LoggingActions';
+import { HandlerAgent } from '../../../agent/HandlerAgent';
+import { ReplyToSkeetAction } from '../../../actions/message-actions/post/SkeetActions';
+import { CreateSkeetMessage } from '../../../types/JetstreamTypes';
 import { CreateSkeetHandler } from '../skeet/CreateSkeetHandler';
 
 // TODO see comment at top of BadBotHandler
