+++ src/validations/message-validators/post/StringValidators.ts
@@ -1,7 +1,7 @@
-import { flattenTextUpdated } from '../../utils/text-utils';
-import { AbstractValidator } from '../AbstractValidator';
-import { HandlerAgent } from '../../agent/HandlerAgent';
-import { CreateSkeetMessage } from '../../types/JetstreamTypes';
+import { flattenTextUpdated } from '../../../utils/text-utils';
+import { AbstractValidator } from '../../AbstractValidator';
+import { HandlerAgent } from '../../../agent/HandlerAgent';
+import { CreateSkeetMessage } from '../../../types/JetstreamTypes';
 
 export class InputIsCommandValidator extends AbstractValidator {
     constructor(
